refactor AppBar into its own component
