+++ src/App.tsx
@@ -1,42 +1,18 @@
-import { AppBar, Badge, CssBaseline, Divider, Drawer, IconButton, makeStyles, Toolbar, Typography, useTheme } from '@material-ui/core'
+import { CssBaseline, Divider, Drawer, IconButton, makeStyles, useTheme } from '@material-ui/core'
 import { HashRouter } from 'react-router-dom'
 import { useState } from 'react'
 import './App.css'
 import clsx from 'clsx'
-import MenuIcon from '@material-ui/icons/Menu'
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
 import ChevronRightIcon from '@material-ui/icons/ChevronRight'
-import GitHubIcon from '@material-ui/icons/GitHub'
-import TwitterIcon from '@material-ui/icons/Twitter'
+import TopAppBar from './components/TopAppBar'
 import NavMenuList from './components/NavMenuList'
 import RouteContainer from './components/RouteContainer'
 
 const drawerWidth = 240
-
 const useStyles = makeStyles((theme) => ({
   root: {
     display: 'flex',
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  menuButton: {
-    marginRight: 36,
-  },
-  hide: {
-    display: 'none',
   },
   drawer: {
     width: drawerWidth,
@@ -72,29 +48,12 @@ const useStyles = makeStyles((theme) => ({
     flexGrow: 1,
     padding: theme.spacing(3),
   },
-  badgeSection: {
-    float: 'right',
-    display: 'none',
-    [theme.breakpoints.up('md')]: {
-      display: 'flex',
-    },
-  },
-  grow: {
-    flexGrow: 1,
-  },
-  toolbarHeader: {
-    fontFamily: '"Courier New",Courier,monospace'
-  }
 }))
 
 function App() {
   const classes = useStyles()
   const theme = useTheme()
   const [open, setOpen] = useState(false)
-
-  const handleDrawerOpen = () => {
-    setOpen(true)
-  }
 
   const handleDrawerClose = () => {
     setOpen(false)
@@ -104,43 +63,7 @@ function App() {
     <div className={classes.root}>
       <HashRouter basename="/">
         <CssBaseline />
-        <AppBar
-          color="default"
-          position="fixed"
-          className={clsx(classes.appBar, {
-            [classes.appBarShift]: open,
-          })}
-        >
-          <Toolbar>
-            <IconButton
-              color="inherit"
-              aria-label="open drawer"
-              onClick={handleDrawerOpen}
-              edge="start"
-              className={clsx(classes.menuButton, {
-                [classes.hide]: open,
-              })}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography variant="h4" className={classes.toolbarHeader} noWrap>
-              mikebents.com
-            </Typography>
-            <div className={classes.grow} />
-            <div className={classes.badgeSection}>
-              <IconButton title="GitHub repos" color="inherit" href="https://github.com/mbents">
-                <Badge color="secondary">
-                  <GitHubIcon />
-                </Badge>
-              </IconButton>
-              <IconButton title="Twitter" color="inherit" href="https://twitter.com/MikeBents">
-                <Badge color="secondary">
-                  <TwitterIcon />
-                </Badge>
-              </IconButton>
-            </div>
-          </Toolbar>
-        </AppBar>
+        <TopAppBar open={open} setOpen={setOpen} />
         <Drawer
           variant="permanent"
           className={clsx(classes.drawer, {
